Order funcionando con customer, falta implementar la relación con order-asset, tambien modifique el nombre de bienDeCambio por asset

+++ back/src/main/java/crisalis/blue/controllers/CustomerController.java
@@ -39,7 +39,7 @@ public class CustomerController {
 
 
     @DeleteMapping(value = "delete")
-    public String deleteCustomer(@RequestParam int id ) {
+    public String deleteCustomer(@RequestParam Long id ) {
         return customerService.deleteCustomer(id);
     }
